Use custom settings.xml for build with release package action
Switch to fork kappmeier/github-actions-maven-release@master to release to
GitHub Packages. The fork allows to provide custom settings.xml file to maven
that contains settings to access GitHub Packagesd during the build phase.

The settings.xml file is built by the whelk-io/maven-settings-xml-action. It
is copied to the working directory and passed to the GitHub Package release
plugin via the variable `settinngs-file`.